backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
Pillow==9.0.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0